gcc-plugins: Add the randstruct plugin
This feature is ported from grsecurity. The implementation is almost
entirely identical to the original code written by the PaX Team and
Brad Spengler.

The new randstruct plugin randomizes the layout of selected structures at
compile time, as a probabilistic defense against attacks that need to
know the layout of structures within the kernel. This is most useful for
"in-house" kernel builds where the neither the randomization seed nor
other build artifacts are made available to an attacker. While less useful
for distribution kernels (where the randomization seed must be exposed for
third party kernel module builds), it still has some value there since now
all kernel builds would need to be tracked by an attacker.

This defines two annotations __randomize_layout and __no_randomize_layout,
which respectively tell the plugin to either randomize or not to randomize
the struct in question. Follow-on patches enable the auto-detection
logic for selecting structures that contain only function pointers. It
is disabled here to assist with bisection.

Since any randomized structs must be initialized using designated
initializers, __randomize_layout includes the __designated_init
annotation even when the plugin is disabled so that all builds will
require the needed initialization. (The plugin adds the annotation
for automatically chosen structures.)

The main differences between this implemenation and grsecurity are:
- disable automatic struct selection (enabled in follow-up patch)
- add designated_init attribute at runtime and for manual marking
- clarify debugging output to differentiate bad cast warnings
- add whitelisting infrastructure
- support gcc 7's DECL_ALIGN and DECL_MODE changes (Laura Abbott)
- raise minimum required GCC version to 4.7

Earlier versions of this patch were ported by Michael Leibowitz.

Signed-off-by: Kees Cook <keescook@chromium.org>