parisc: Stop CPUs via PAT firmware before system halt or reboot.
Dave reported that he had issued a "shutdown -r" and a panic occurred during
the reboot while all CPUs were still up. After this, stall messages were output
to console after the firmware version was printed.

To avoid that issue, add functions to call PAT firmware to stop all CPUs (with
the exception of the currently running CPU) before a panic reboot or a system
halt is issued.

Signed-off-by: Helge Deller <deller@gmx.de>